rename Lannion 2022
