+++ api/src/pdc/services/policy/engine/policies/20230501_Lannion.spec.ts
@@ -2,7 +2,7 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { Lannion as Handler } from './Lannion';
+import { Lannion as Handler } from './20230501_Lannion';
 
 const defaultPosition = {
   arr: '22113',
